+++ stroom-explorer-server/src/main/java/stroom/explorer/server/ExplorerTreeModel.java
@@ -133,4 +133,8 @@ class ExplorerTreeModel {
 
         explorerSession.setMinExplorerTreeModelBuildTime(now);
     }
+
+    void clear() {
+        currentModel = null;
+    }
 }