+++ stroom-security-server/src/main/java/stroom/security/server/UserDocumentPermissions.java
@@ -1,0 +1,5 @@
+package stroom.security.server;
+
+public interface UserDocumentPermissions {
+    boolean hasDocumentPermission(String docUuid, String permission);
+}