+++ stroom-security-server/src/main/java/stroom/security/server/UserDocumentPermissionsCache.java
@@ -26,51 +26,56 @@ import stroom.entity.server.event.EntityEventHandler;
 import stroom.entity.shared.EntityAction;
 import stroom.node.server.StroomPropertyService;
 import stroom.query.api.v2.DocRef;
-import stroom.security.shared.DocumentPermissions;
+import stroom.security.shared.UserRef;
 import stroom.util.cache.CacheManager;
 import stroom.util.cache.CacheUtil;
 
 import javax.inject.Inject;
 import javax.inject.Provider;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
 import java.util.concurrent.TimeUnit;
+import java.util.stream.Collectors;
 
 @Component
 @EntityEventHandler(action = EntityAction.CLEAR_CACHE)
-public class DocumentPermissionsCache implements EntityEvent.Handler {
-    public static final String MAXIMUM_SIZE_PROPERTY = "stroom.security.documentPermissions.maxCacheSize";
+public class UserDocumentPermissionsCache implements EntityEvent.Handler {
+    private static final String MAXIMUM_SIZE_PROPERTY = "stroom.security.documentPermissions.maxCacheSize";
 
     private static final int DEFAULT_MAXIMUM_SIZE = 100000;
 
     private final Provider<EntityEventBus> eventBusProvider;
     private final StroomPropertyService stroomPropertyService;
+    private final UserGroupsCache userGroupsCache;
 
     private final CacheManager cacheManager;
     private final DocumentPermissionService documentPermissionService;
 
     private volatile Integer lastMaximumSize;
-    private volatile LoadingCache<DocRef, DocumentPermissions> cache;
+    private volatile LoadingCache<UserRef, UserDocumentPermissions> cache;
 
     @Inject
-    public DocumentPermissionsCache(final CacheManager cacheManager,
-                                    final DocumentPermissionService documentPermissionService,
-                                    final Provider<EntityEventBus> eventBusProvider,
-                                    final StroomPropertyService stroomPropertyService) {
+    public UserDocumentPermissionsCache(final CacheManager cacheManager,
+                                        final DocumentPermissionService documentPermissionService,
+                                        final Provider<EntityEventBus> eventBusProvider,
+                                        final StroomPropertyService stroomPropertyService,
+                                        final UserGroupsCache userGroupsCache) {
         this.cacheManager = cacheManager;
         this.documentPermissionService = documentPermissionService;
         this.eventBusProvider = eventBusProvider;
         this.stroomPropertyService = stroomPropertyService;
+        this.userGroupsCache = userGroupsCache;
     }
 
-    DocumentPermissions get(final DocRef key) {
-        return getCache().getUnchecked(key);
+    UserDocumentPermissions get(final UserRef userRef) {
+        return getCache().getUnchecked(userRef);
     }
 
-    void remove(final DocRef docRef) {
-        if (cache != null) {
-            cache.invalidate(docRef);
-        }
+    void removeAll() {
+        clear();
         final EntityEventBus entityEventBus = eventBusProvider.get();
-        EntityEvent.fire(entityEventBus, docRef, EntityAction.CLEAR_CACHE);
+        EntityEvent.fire(entityEventBus, null, EntityAction.CLEAR_CACHE);
     }
 
     void clear() {
@@ -82,7 +87,8 @@ public class DocumentPermissionsCache implements EntityEvent.Handler {
     @Override
     public void onChange(final EntityEvent event) {
         if (cache != null) {
-            cache.invalidate(event.getDocRef());
+            clear();
+//            cache.invalidate(event.getDocRef());
         }
     }
 
@@ -90,7 +96,7 @@ public class DocumentPermissionsCache implements EntityEvent.Handler {
         return stroomPropertyService.getIntProperty(MAXIMUM_SIZE_PROPERTY, DEFAULT_MAXIMUM_SIZE);
     }
 
-    private LoadingCache<DocRef, DocumentPermissions> getCache() {
+    private LoadingCache<UserRef, UserDocumentPermissions> getCache() {
         if (lastMaximumSize == null || lastMaximumSize != getMaximumSize()) {
             createCache();
         }
@@ -101,15 +107,21 @@ public class DocumentPermissionsCache implements EntityEvent.Handler {
     private synchronized void createCache() {
         final int maximumSize = getMaximumSize();
         if (lastMaximumSize == null || lastMaximumSize != maximumSize) {
-            final CacheLoader<DocRef, DocumentPermissions> cacheLoader = CacheLoader.from(documentPermissionService::getPermissionsForDocument);
+            final CacheLoader<UserRef, UserDocumentPermissions> cacheLoader = CacheLoader.from(userRef -> {
+                final List<UserRef> userGroups = userGroupsCache.get(userRef);
+                final Set<String> users = new HashSet<>();
+                users.add(userRef.getUuid());
+                users.addAll(userGroups.stream().map(DocRef::getUuid).collect(Collectors.toSet()));
+                return documentPermissionService.getPermissionsForUsers(users);
+            });
             final CacheBuilder cacheBuilder = CacheBuilder.newBuilder()
                     .maximumSize(maximumSize)
-                    .expireAfterAccess(30, TimeUnit.MINUTES);
-            final LoadingCache<DocRef, DocumentPermissions> cache = cacheBuilder.build(cacheLoader);
+                    .expireAfterAccess(10, TimeUnit.MINUTES);
+            final LoadingCache<UserRef, UserDocumentPermissions> cache = cacheBuilder.build(cacheLoader);
             if (lastMaximumSize == null) {
-                cacheManager.registerCache("Document Permissions Cache", cacheBuilder, cache);
+                cacheManager.registerCache("User Document Permissions Cache", cacheBuilder, cache);
             } else {
-                cacheManager.replaceCache("Document Permissions Cache", cacheBuilder, cache);
+                cacheManager.replaceCache("User Document Permissions Cache", cacheBuilder, cache);
             }
             lastMaximumSize = maximumSize;
             this.cache = cache;